employee administration

+++ employees.php
@@ -1,8 +1,8 @@
 <?
-include 'pages/includes/header.php';
+include 'includes/header.php';
 
 if (!isset($_SESSION['user'])) {
-    if ($_SESSION['user']['roleId'] != 11) {
+    if ($_SESSION['user']['roleId'] != 1) {
         header("Location: login.php");
         exit;
     }
@@ -13,9 +13,13 @@ require('classes/User.php');
 $user = new User();
 $employees = $user->getEmployees();
 ?>
-
     <div class="container">
+    <br>
+    <br>
         <!-- Example row of columns -->
+        <div class="row">
+            <a href="add_employee.php" class="btn btn-default">Add Employee</a>
+        </div>
         <div class="row">
             <br>
             <br>
@@ -28,27 +32,32 @@ $employees = $user->getEmployees();
                     <th>Action</th>
                 </thead>
                 <tbody>
-                    <? foreach ($employees as $employee): ?>
-                        <td><?=  $employee['firstName'] ?></td>
-                        <td><?=  $employee['lastName'] ?></td>
-                        <td><?=  $employee['email'] ?></td>
-                        <td><?=  $employee['daysLeft'] ?></td>
-                        <td>
-                            <a href="edit_employee.php?id=<?=  $employee['id']?>">Edit</a>
-                            &nbsp;&nbsp;
-                            <a href="delete_employee.php?id=<?=  $employee['id']?>">Delete</a>
-                        </td>
-                    <? endforeach; ?>
+                    
+                    <?
+                    if ($employees):
+                        foreach ($employees as $employee): ?>
+                            <tr>
+                                <td><?=  $employee['firstName'] ?></td>
+                                <td><?=  $employee['lastName'] ?></td>
+                                <td><?=  $employee['email'] ?></td>
+                                <td><?=  $employee['daysLeft'] ?></td>
+                                <td>
+                                    <a href="edit_employee.php?id=<?=  $employee['id']?>">Edit</a>
+                                    &nbsp;&nbsp;
+                                    <a href="delete_employee.php?id=<?=  $employee['id']?>">Delete</a>
+                                </td>
+                            </tr>
+                    <?  endforeach; 
+                    endif;
+                    ?>
                 </tbody>
             </table>
         </div>
-
-
 <?
-include 'pages/includes/footer.php';
+include 'includes/footer.php';
 ?>
-    <script type="application/javascript">
-        $(document).ready(function () {
-            $('#dataTable').dataTable();
-        })
-    </script>
+<script type="application/javascript">
+    $(document).ready(function () {
+        $('#dataTable').dataTable();
+    })
+</script>
